Adding constructor for full circular layer and changing return type on getReinfBarDiameter()
git-svn-id: svn://peera.berkeley.edu/usr/local/svn/OpenSees/trunk@445 0eff8f49-d3cf-49fd-9ff1-3c36910be21e

+++ SRC/material/section/repres/reinfLayer/ReinfLayer.h
@@ -18,8 +18,8 @@
 **                                                                    **
 ** ****************************************************************** */
                                                                         
-// $Revision: 1.1.1.1 $
-// $Date: 2000-09-15 08:23:22 $
+// $Revision: 1.2 $
+// $Date: 2001-08-10 18:00:38 $
 // $Source: /usr/local/cvs/OpenSees/SRC/material/section/repres/reinfLayer/ReinfLayer.h,v $
                                                                         
                                                                         
@@ -52,7 +52,7 @@ class ReinfLayer
     
     virtual int         getNumReinfBars     (void) const = 0;
     virtual int         getMaterialID       (void) const = 0; 
-    virtual int         getReinfBarDiameter (void) const = 0;
+    virtual double      getReinfBarDiameter (void) const = 0;
     virtual double      getReinfBarArea     (void) const = 0;
     virtual ReinfLayer *getCopy             (void) const = 0;
     virtual ReinfBar   *getReinfBars        (void) const = 0;
